Add support for non-square images
The widget is limited by the fact that the width and height of the star
are expected to be the same. The image doesn't actually have to be a
star and could have a rectangular shape.
Note that this change is not backward compatible with previous version.
Also I didn't rename the term "star" to something more generic as
part of this change but it would make sense to.